requirements.txt
streamlit>=1.28.0
openai>=1.0.0
pymupdf>=1.23.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
Pillow>=10.0.0
python-dotenv>=1.0.0
numpy>=1.26.0